requirements.txt
Flask==3.0.2
pymongo==4.6.2
sentence-transformers==2.5.1
faiss-cpu==1.8.0
rank-bm25==0.2.2
numpy==1.26.4
scipy==1.12.0
nltk==3.8.1
python-dotenv==1.0.1
torch==2.2.1
transformers==4.38.2
PyPDF2==3.0.1
python-pptx==0.6.23
